Split the amx_api.c to create a regression test
